Update reference-connect-ports.md
During working on SR 121030324001275 while troubleshooting we found out that AD Connect having it's DNS set to a Primary Domain and this domain have a secondary zone for a second separate domain. While trying to add this new second domain to AD Connect, when configuring SSO for it, the message: getdesktopssocomputeraccountdns exception caught: the server is not operational was recorded at the trace log files. After capturing a network trace, we found that AD Connect tries to call the global catalog for the second domain on port 3268. This is not documented, hence the changed proposed here.